Dodan prevod za drugu lekciju
Potrebno dodati prevod i za zoom pregled!
